getOrders ok

+++ src/database/models/order.module.ts
@@ -1,17 +1,16 @@
 import { sequelize } from "../pg";
 import { DataTypes } from "sequelize";
-export const orderModel = sequelize.define(
-  "orders",
-  {
-    orderId: {
-      type: DataTypes.INTEGER,
-      primaryKey: true,
-      autoIncrement: true,
-    },
-    total: {
-      type: DataTypes.DECIMAL,
-    },
-    productId: {
-      type: DataTypes.DECIMAL,
-    },
-  });
+export const orderModel = sequelize.define("orders", {
+  orderId: {
+    type: DataTypes.INTEGER,
+    primaryKey: true,
+    autoIncrement: true,
+  },
+  total: {
+    type: DataTypes.INTEGER,
+  },
+  productId: {
+    type: DataTypes.INTEGER,
+  },
+});
+
